requirements.txt
Django==2.2.5
django-environ==0.4.5
djangorestframework==3.10.3
entrypoints==0.3
mccabe==0.6.1
pytz==2019.2
sqlparse==0.3.0
